Barcode upc is bigint

+++ db/migrate/20150706104432_create_barcodes.rb
@@ -1,8 +1,7 @@
 class CreateBarcodes < ActiveRecord::Migration
   def change
     create_table :barcodes do |t|
-      t.integer :upc
-
+      t.bigint :upc
       t.timestamps null: false
     end
   end
